chore: new package for payment detection (#126)
