+++ .circleci/config.yml
@@ -420,6 +420,30 @@ jobs:
           root: *working_directory
           paths:
             - packages/multi-format/coverage/
+  test-payment-detection:
+    docker:
+      - *node_image
+      - *ganache_image
+    working_directory: *working_directory
+    steps:
+      - attach_workspace:
+          at: *working_directory
+      - run:
+          name: 'Build payment-detection'
+          command: 'yarn workspace @requestnetwork/payment-detection run build'
+      - run:
+          name: 'Lint payment-detection'
+          command: 'yarn workspace @requestnetwork/payment-detection run lint'
+      - run:
+          name: 'Deploy test contracts from smart-contracts'
+          command: 'yarn workspace @requestnetwork/smart-contracts run deploy'
+      - run:
+          name: 'Test payment-detection'
+          command: 'yarn workspace @requestnetwork/payment-detection run test'
+      - persist_to_workspace:
+          root: *working_directory
+          paths:
+            - packages/payment-detection/coverage/
   publish-coverage:
     docker:
       - *node_image
@@ -590,6 +614,9 @@ workflows:
       - test-prototype-estimator:
           requires:
             - build
+      - test-payment-detection:
+          requires:
+            - build
 
       # Publish the code coverage
       - publish-coverage:
@@ -607,6 +634,7 @@ workflows:
             - test-request-node
             - test-utils
             - test-multi-format
+            - test-payment-detection
 
       # Release a next version package everytime we merge to master
       - next-release:
@@ -627,6 +655,7 @@ workflows:
             - test-types
             - test-utils
             - test-web3-signature
+            - test-payment-detection
           context: protocol-release
           filters:
             branches:
