+++ packages/payment-detection/.lintstagedrc.json
@@ -1,0 +1,3 @@
+{
+  "src/**/*.ts": ["tslint --project . --fix", "prettier --single-quote --write", "git add"]
+}
