+++ packages/payment-detection/package.json
@@ -1,0 +1,72 @@
+{
+  "name": "@requestnetwork/payment-detection",
+  "version": "0.12.0",
+  "publishConfig": {
+    "access": "public"
+  },
+  "description": "Payment detection using ethers.",
+  "keywords": [
+    "requestnetwork"
+  ],
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/RequestNetwork/requestNetwork.git"
+  },
+  "homepage": "https://github.com/RequestNetwork/requestNetwork/tree/master/packages/payment-detection#readme",
+  "bugs": {
+    "url": "https://github.com/RequestNetwork/requestNetwork/issues"
+  },
+  "license": "MIT",
+  "engines": {
+    "node": ">=8.0.0"
+  },
+  "main": "dist/index.js",
+  "types": "dist/index.d.ts",
+  "directories": {
+    "lib": "src",
+    "test": "test"
+  },
+  "files": [
+    "dist"
+  ],
+  "scripts": {
+    "build": "tsc -b",
+    "clean": "shx rm -rf dist tsconfig.tsbuildinfo",
+    "lint": "tslint --project . && eslint \"src/**/*.ts\"",
+    "lint-staged": "lint-staged",
+    "prepare": "yarn run build",
+    "test": "nyc mocha --extension ts --require ts-node/register --require source-map-support/register \"test/**/*.ts\"",
+    "test:watch": "yarn test --watch"
+  },
+  "dependencies": {
+    "@requestnetwork/types": "0.11.0",
+    "@requestnetwork/utils": "0.10.0",
+    "axios": "0.19.0",
+    "bn.js": "4.11.8",
+    "ethers": "4.0.43",
+    "node-fetch": "2.6.0",
+    "satoshi-bitcoin": "1.0.4"
+  },
+  "devDependencies": {
+    "@types/bn.js": "4.11.5",
+    "@types/chai": "4.1.7",
+    "@types/mocha": "5.2.7",
+    "@types/sinon": "7.5.0",
+    "@typescript-eslint/parser": "1.2.0",
+    "chai": "4.2.0",
+    "chai-as-promised": "7.1.1",
+    "chai-spies": "1.0.0",
+    "eslint": "5.13.0",
+    "eslint-plugin-spellcheck": "0.0.14",
+    "eslint-plugin-typescript": "0.14.0",
+    "lint-staged": "8.1.3",
+    "mocha": "6.2.2",
+    "nyc": "15.0.0",
+    "prettier": "1.16.4",
+    "shx": "0.3.2",
+    "sinon": "7.5.0",
+    "source-map-support": "0.5.13",
+    "tslint": "5.12.1",
+    "typescript": "3.7.2"
+  }
+}
