+++ packages/payment-detection/src/btc/default-bitcoin-detection-provider.ts
