+++ packages/payment-detection/src/declarative.ts
