+++ packages/payment-detection/src/erc20/index.ts
@@ -1,0 +1,4 @@
+import AddressBased from './address-based';
+import ProxyContract from './proxy-contract';
+
+export { AddressBased, ProxyContract };
