+++ packages/payment-detection/src/eth/index.ts
@@ -1,0 +1,3 @@
+import InputData from './input-data';
+
+export { InputData };
