+++ packages/payment-detection/src/index.ts
@@ -1,0 +1,16 @@
+import PaymentNetworkFactory from './payment-network-factory';
+import PaymentReferenceCalculator from './payment-reference-calculator';
+
+import * as BtcPaymentNetwork from './btc';
+import DeclarativePaymentNetwork from './declarative';
+import * as Erc20PaymentNetwork from './erc20';
+import * as EthPaymentNetwork from './eth';
+
+export {
+  PaymentNetworkFactory,
+  PaymentReferenceCalculator,
+  BtcPaymentNetwork,
+  DeclarativePaymentNetwork,
+  Erc20PaymentNetwork,
+  EthPaymentNetwork,
+};
