+++ packages/payment-detection/src/payment-reference-calculator.ts
