+++ packages/payment-detection/test/btc/default-providers/blockstream-info.test.ts
@@ -1,7 +1,8 @@
 import { PaymentTypes } from '@requestnetwork/types';
-import BlockchainInfo from '../../../../../src/api/payment-network/btc/default-providers/blockchain-info';
 
-import * as BlockchainInfoData from './blockchain-info-data';
+import Blockstream from '../../../src/btc/default-providers/blockstream-info';
+
+import * as BlockstreamData from './blockstream-info-data';
 
 import * as chai from 'chai';
 import * as chaiAsPromised from 'chai-as-promised';
@@ -13,12 +14,16 @@ import 'mocha';
 
 // Most of the tests are done as integration tests in ../index.test.ts
 /* tslint:disable:no-unused-expression */
-describe('api/btc/default-providers/blockchainInfo', () => {
+describe('api/btc/providers/blockstream', () => {
   describe('getAddressInfo', () => {
     it('must throw if bitcoinNetworkId is not 0 or 3', async () => {
-      const blockchainInfo = new BlockchainInfo();
+      const blockstreamData = new Blockstream();
       expect(
-        blockchainInfo.getAddressBalanceWithEvents(1, 'address', PaymentTypes.EVENTS_NAMES.PAYMENT),
+        blockstreamData.getAddressBalanceWithEvents(
+          1,
+          'address',
+          PaymentTypes.EVENTS_NAMES.PAYMENT,
+        ),
       ).to.eventually.be.rejectedWith(
         'Invalid network 0 (mainnet) or 3 (testnet) was expected but 1 was given',
       );
@@ -27,14 +32,14 @@ describe('api/btc/default-providers/blockchainInfo', () => {
 
   describe('parse', () => {
     it('can parse data', () => {
-      const blockchainInfo = new BlockchainInfo();
-
-      const parsedData = blockchainInfo.parse(
-        BlockchainInfoData.exampleAddressInfo,
+      const blockstreamData = new Blockstream();
+      const parsedData = blockstreamData.parse(
+        { txs: BlockstreamData.exampleAddressInfo, address: 'mgPKDuVmuS9oeE2D9VPiCQriyU14wxWS1v' },
         PaymentTypes.EVENTS_NAMES.PAYMENT,
       );
       expect(parsedData.balance, 'balance wrong').to.equal('50500000');
-      expect(parsedData.events, 'balance wrong').to.deep.equal([
+
+      expect(parsedData.events, 'events wrong').to.deep.equal([
         {
           amount: '500000',
           name: 'payment',
@@ -42,7 +47,7 @@ describe('api/btc/default-providers/blockchainInfo', () => {
             block: 1354204,
             txHash: '2a14f1ad2dfa4601bdc7a6be325241bbdc2ae99d05f096357fda76264b1c5c26',
           },
-          timestamp: 1531879904,
+          timestamp: 1531880048,
         },
         {
           amount: '50000000',
@@ -51,7 +56,7 @@ describe('api/btc/default-providers/blockchainInfo', () => {
             block: 1354075,
             txHash: '7d84924c034798dedcc95f479c9cdb24fe014437f7ce0ee0c2f4bf3580e017d8',
           },
-          timestamp: 1531817766,
+          timestamp: 1531818367,
         },
       ]);
     });
