+++ packages/payment-detection/test/payment-network-factory.test.ts
@@ -1,15 +1,18 @@
-import { AdvancedLogicTypes, ExtensionTypes, RequestLogicTypes } from '@requestnetwork/types';
+import {
+  AdvancedLogicTypes,
+  ExtensionTypes,
+  PaymentTypes,
+  RequestLogicTypes,
+} from '@requestnetwork/types';
 
-import BTCAddressedBased from '../../../src/api/payment-network/btc/mainnet-address-based';
-import Declarative from '../../../src/api/payment-network/declarative';
+import BTCAddressedBased from '../src/btc/mainnet-address-based';
+import Declarative from '../src/declarative';
 
 import { expect } from 'chai';
 
 import 'mocha';
 
-import PaymentNetworkFactory from '../../../src/api/payment-network/payment-network-factory';
-
-import { PaymentTypes } from '@requestnetwork/types';
+import PaymentNetworkFactory from '../src/payment-network-factory';
 
 const mockAdvancedLogic: AdvancedLogicTypes.IAdvancedLogic = {
   applyActionToExtensions(): any {
