+++ packages/request-client.js/package.json
@@ -48,6 +48,7 @@
     "@requestnetwork/data-format": "0.4.8",
     "@requestnetwork/epk-signature": "0.5.7",
     "@requestnetwork/multi-format": "0.2.4",
+    "@requestnetwork/payment-detection": "0.12.0",
     "@requestnetwork/request-logic": "0.8.3",
     "@requestnetwork/smart-contracts": "0.3.0",
     "@requestnetwork/transaction-manager": "0.8.4",
@@ -57,9 +58,7 @@
     "bn.js": "4.11.8",
     "currency-codes": "1.5.1",
     "eth-contract-metadata": "1.11.0",
-    "ethers": "4.0.43",
-    "node-fetch": "2.3.0",
-    "satoshi-bitcoin": "1.0.4"
+    "ethers": "4.0.43"
   },
   "devDependencies": {
     "@compodoc/compodoc": "1.1.10",
