+++ packages/request-client.js/src/index.ts
@@ -1,5 +1,6 @@
+import { PaymentReferenceCalculator } from '@requestnetwork/payment-detection';
+
 import Currency from './api/currency';
-import PaymentReferenceCalculator from './api/payment-network/payment-reference-calculator';
 import Request from './api/request';
 import Utils from './api/utils';
 import { default as RequestNetwork } from './http-request-network';
